+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 <!--
     ## **WORK IN PROGRESS**
 -->
+## **WORK IN PROGRESS**
+
 ## 0.4.3 (2023-01-31)
 - Fixed Switchlight when lightpath is empty